feat(widevine): change base64dec api

+++ framework/codec/Android/mediacodec_jni.cpp
@@ -23,8 +23,8 @@ namespace Cicada {
             mediaDrm  = std::unique_ptr<MediaDrmWrapper>(new MediaDrmWrapper("edef8ba9-79d6-4ace-a3c8-27dcd51d21ed"));
             std::map<std::string, std::string> param{};
             char *dst = nullptr;
-            uint64_t len  = CicadaUtils::base64dec("AAAARHBzc2gAAAAA7e+LqXnWSs6jyCfc1R0h7QAAACQIARIBNRoNd2lkZXZpbmVfdGVzdCIKMjAxNV90ZWFycyoCU0Q=",
-                                                   &dst);
+            int len  = CicadaUtils::base64dec("AAAARHBzc2gAAAAA7e+LqXnWSs6jyCfc1R0h7QAAACQIARIBNRoNd2lkZXZpbmVfdGVzdCIKMjAxNV90ZWFycyoCU0Q=",
+                                              &dst);
             mediaDrm->setKeyRequestInfo("https://proxy.uat.widevine.com/proxy?provider=widevine_test",
                                         dst, len, "video/mp4",
                                         1 /*MediaDrm.KEY_TYPE_STREAMING*/,
